+++ beforeido/Home 2.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+
+const Home = () => {
+  return (
+    <View>
+      <Text>Welcome to the Home Page!</Text>
+    </View>
+  );
+};
+
+export default Home;
