+++ beforeido/screens/ConsultationScreen.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { View, Text, Button } from 'react-native';
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { RootStackParamList } from '../App'; // Import RootStackParamList from App
+
+type ConsultationScreenProps = NativeStackScreenProps<RootStackParamList, "Consultation">;
+
+const ConsultationScreen: React.FC<ConsultationScreenProps> = (props) => {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Consultation Screen</Text>
+      <Button
+        title="Go to Home"
+        onPress={() => props.navigation.push('Home')}
+      />
+    </View>
+  );
+};
+
+export default ConsultationScreen;
