+++ beforeido/screens/HomeScreen.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { View, Text, Button } from 'react-native';
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { RootStackParamList } from '../App'; // Import RootStackParamList from App
+
+type HomeScreenProps = NativeStackScreenProps<RootStackParamList, "Home">;
+
+const HomeScreen: React.FC<HomeScreenProps> = (props) => {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Home Screen</Text>
+      <Button
+        title="Go to Profile"
+        onPress={() => props.navigation.push('Profile')}
+      />
+      <Button
+        title="Take the Assessment"
+        onPress={() => props.navigation.push('Assessment')}
+      />
+      <Button
+        title="View Results"
+        onPress={() => props.navigation.push('Results')}
+      />
+      <Button
+        title="Consultation"
+        onPress={() => props.navigation.push('Consultation')}
+      />
+      <Text>Your Couple Code: 20</Text>
+    </View>
+  );
+};
+
+export default HomeScreen;
