+++ beforeido/screens/LoginScreen.tsx
@@ -1,0 +1,59 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, Button } from 'react-native';
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { RootStackParamList } from '../App'; // Import RootStackParamList from App
+
+type LoginScreenProps = NativeStackScreenProps<RootStackParamList, "Login">;
+
+const LoginScreen: React.FC<LoginScreenProps> = (props) => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [coupleCode, setCoupleCode] = useState('');
+
+  const handleLogin = () => {
+    console.log('Logging in...');
+    console.log('Email:', email);
+    console.log('Password:', password);
+    console.log('Couple Code:', coupleCode);
+    props.navigation.push('Home');
+  };
+
+  const handleSignUp = () => {
+    props.navigation.push('Signup');
+  };
+  
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Login Screen</Text>
+      <TextInput
+        placeholder="Email"
+        value={email}
+        onChangeText={setEmail}
+        style={{ height: 40, width: 300, borderColor: 'gray', borderWidth: 1, marginVertical: 10 }}
+      />
+      <TextInput
+        placeholder="Password"
+        value={password}
+        onChangeText={setPassword}
+        secureTextEntry={true}
+        style={{ height: 40, width: 300, borderColor: 'gray', borderWidth: 1, marginVertical: 10 }}
+      />
+      <TextInput
+        placeholder="Couple Code"
+        value={coupleCode}
+        onChangeText={setCoupleCode}
+        style={{ height: 40, width: 300, borderColor: 'gray', borderWidth: 1, marginVertical: 10 }}
+      />
+      <Button
+        title="Login"
+        onPress={handleLogin}
+      />
+      <Button
+        title="Sign Up"
+        onPress={handleSignUp}
+      />
+    </View>
+  );
+};
+
+export default LoginScreen;
