+++ beforeido/screens/ProfileScreen.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import { View, Text, Button } from 'react-native';
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { RootStackParamList } from '../App'; // Import RootStackParamList from App
+
+type ProfileScreenProps = NativeStackScreenProps<RootStackParamList, "Profile">;
+
+const ProfileScreen: React.FC<ProfileScreenProps> = (props) => {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Profile Screen</Text>
+      <Button
+        title="Go to Home"
+        onPress={() => props.navigation.push('Home')}
+      />
+    </View>
+  );
+};
+
+export default ProfileScreen;
