+++ beforeido/screens/SignupScreen.tsx
@@ -1,0 +1,53 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, Button } from 'react-native';
+import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { RootStackParamList } from '../App'; // Import RootStackParamList from App
+
+type SignupScreenProps = NativeStackScreenProps<RootStackParamList, "Signup">;
+
+const SignupScreen: React.FC<SignupScreenProps> = (props) => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [coupleCode, setCoupleCode] = useState('');
+
+  const handleSignUp = () => {
+    // Here you can implement your signup logic
+    console.log('Signing up...');
+    console.log('Email:', email);
+    console.log('Password:', password);
+    console.log('Couple Code:', coupleCode);
+    // After successful login, you can navigate to the login screen
+    props.navigation.push('Login');
+  };
+
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Sign Up Screen</Text>
+      <TextInput
+        placeholder="Email"
+        value={email}
+        onChangeText={setEmail}
+        style={{ height: 40, width: 300, borderColor: 'gray', borderWidth: 1, marginVertical: 10 }}
+      />
+      <TextInput
+        placeholder="Password"
+        value={password}
+        onChangeText={setPassword}
+        secureTextEntry={true}
+        style={{ height: 40, width: 300, borderColor: 'gray', borderWidth: 1, marginVertical: 10 }}
+      />
+      <TextInput
+        placeholder="Couple Code - Optional"
+        value={coupleCode}
+        onChangeText={setCoupleCode}
+        style={{ height: 40, width: 300, borderColor: 'gray', borderWidth: 1, marginVertical: 10 }}
+      />
+      <Button
+        title="Sign Up"
+        onPress={handleSignUp}
+      />
+    </View>
+  );
+};
+
+export default SignupScreen;
